Modify to pass compile

+++ 0001-0050/prob0007.rs
@@ -3,5 +3,5 @@ use euler;
 import euler::prime;
 
 fn main() {
-    io::println(#fmt("%u", prime::prime()[10000u64]));
+    io::println(#fmt("%u", prime::prime()[10000u] as uint));
 }
