+++ 0001-0050/prob0012.rs
@@ -15,10 +15,10 @@ fn each_triangles(f: fn(&&uint) -> bool) {
 }
 
 fn num_factors(num: u64, primes: prime::prime) -> u64 {
-    let mut prod = 1u;
+    let mut prod = 1u64;
     for prime::factors(num, primes) {|f|
         let (_base, exp): (u64, i64) = f;
-        prod *= ((exp + 1) as u64);
+        prod *= ((exp + 1i64) as u64);
     }
     ret prod;
 }
@@ -26,9 +26,9 @@ fn num_factors(num: u64, primes: prime::prime) -> u64 {
 fn main() {
     let primes = prime::prime();
     for each_triangles {|t|
-        let num = num_factors(t, primes);
-        if num > 500u {
-            io::println(#fmt("%u -> %u", t, num_factors(t, primes)));
+        let num = num_factors(t as u64, primes);
+        if num > 500u64 {
+            io::println(#fmt("%u -> %u", t, num_factors(t as u64, primes) as uint));
             break;
         }
     }
